requirements.txt
streamlit
pandas
requests
python-dotenv
plotly
numpy
scipy
yfinance
pybaseball
beautifulsoup4
lxml
